bugfix: ScriptChunk.decodeOpN should work for OP_0
OP_0 is a funny opcode. According to the Script class, it `isPushData` and is not `isOpCode`. Yet, like the OP_N opcodes, which all `isOpCode`, it isn't followed by any data, because the data that it pushes is _always_ empty. The OP_N decoding in Script is happy to decode it. ScriptChunk should behave consistently.